Test réglage controleurs
